+++ examples/access-control-migration/src/components/casbin-authorization/index.ts
@@ -1,0 +1,2 @@
+export * from './casbin-authorization-component';
+export * from './services';
