+++ examples/access-control-migration/src/components/casbin-authorization/services/index.ts
@@ -1,0 +1,3 @@
+export * from './assign-project-instance-id.voter';
+export * from './casbin.authorizer';
+export * from './casbin.enforcers';
