+++ examples/access-control-migration/src/components/jwt-authentication/index.ts
@@ -1,0 +1,3 @@
+export * from './jwt-authentication-component';
+export * from './keys';
+export * from './services';
