+++ examples/access-control-migration/src/components/jwt-authentication/services/index.ts
@@ -1,0 +1,4 @@
+export * from './jwt.auth.strategy';
+export * from './jwt.service';
+export * from './security.spec';
+export * from './user.service';
