+++ examples/access-control-migration/src/components/jwt-authentication/services/jwt.service.ts
@@ -8,7 +8,7 @@ import {inject} from '@loopback/context';
 import {HttpErrors} from '@loopback/rest';
 import {securityId, UserProfile} from '@loopback/security';
 import {promisify} from 'util';
-import {TokenServiceBindings} from './keys';
+import {TokenServiceBindings} from '../keys';
 
 const jwt = require('jsonwebtoken');
 const signAsync = promisify(jwt.sign);
