+++ examples/access-control-migration/src/components/jwt-authentication/services/user.service.ts
@@ -8,8 +8,8 @@ import {repository} from '@loopback/repository';
 import {HttpErrors} from '@loopback/rest';
 import {securityId, UserProfile} from '@loopback/security';
 import {compare} from 'bcryptjs';
-import {User} from '../models/user.model';
-import {UserRepository} from '../repositories/user.repository';
+import {User} from '../../../models/user.model';
+import {UserRepository} from '../../../repositories/user.repository';
 
 /**
  * A pre-defined type for user credentials. It assumes a user logs in
