--- examples/access-control-migration/src/services/index.ts
@@ -1,4 +1,0 @@
-// Copyright IBM Corp. 2020. All Rights Reserved.
-// Node module: @loopback/example-access-control-migration
-// This file is licensed under the MIT License.
-// License text available at https://opensource.org/licenses/MIT
